requirements.txt
asttokens==2.4.1
asyncio==3.4.3
backcall==0.2.0
bleach==6.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
executing==2.0.1
Flask==3.0.0
gym==0.26.2
gym-notices==0.0.8
heic-to-jpg==0.1.9
idna==3.6
ipykernel==6.26.0
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.5.0
jupyter_core==5.4.0
kaggle==1.5.16
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
parso==0.8.3
pickleshare==0.7.5
Pillow==10.1.0
pillow-heif==0.13.1
platformdirs==3.11.0
poke-env==0.8.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
py-desmume==0.0.6
pyboy==1.6.10
Pygments==2.16.1
pymongo==4.6.1
PySDL2==0.9.16
pysdl2-dll==2.28.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.3
tabulate==0.9.0
text-unidecode==1.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.12.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.8
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
